Dictionary
Dictionary is a program which enables to review certain words and phrases in Hungarian and in Polish languages. The words and phrases are saved into a MySQL database and the program randomly takes the words from there to review them. It is possible to select the number of words that user wants to review and afterwards percentage of correct answers is calculated. It is also possible to practice on two songs (one Hungarian, one Polish) by adding the randomly removed words and program will check if the answer is correct.